Fix incorrect 'You told <player>' message with 'tell' command.
When you issued a 'tell' command, the message you got was incorrect.  For
example, if players named 'Nick' and 'Test' are playing, and Nick issues
the comand

    tell test hello, tinymudserver world!

Then, Nick would see the following message:

This is obviously incorect.  This commit fixes the commands/DoTell
function to display the correct message

    You tell Test, "hello, tinymudserver world!"